refactor(payments-paypal): Move getPayPalAmountStringFromAmountInCents to a util
