+++ libs/payments/paypal/src/lib/paypal.manager.ts
@@ -8,7 +8,6 @@ import { SubscriptionManager } from '@fxa/payments/stripe';
 import { PayPalClient } from './paypal.client';
 import { BillingAgreement, BillingAgreementStatus } from './paypal.types';
 import { PaypalCustomerMultipleRecordsError } from './paypalCustomer/paypalCustomer.error';
-import { AmountExceedsPayPalCharLimitError } from './paypal.error';
 import { PaypalCustomerManager } from './paypalCustomer/paypalCustomer.manager';
 import { PaypalManagerError } from './paypal.error';
 
@@ -115,23 +114,5 @@ export class PayPalManager {
       throw new PaypalCustomerMultipleRecordsError(uid);
 
     return firstRecord.billingAgreementId;
-  }
-
-  /*
-   * Convert amount in cents to paypal AMT string.
-   * We use Stripe to manage everything and plans are recorded in an AmountInCents.
-   * PayPal AMT field requires a string of 10 characters or less, as documented here:
-   * https://developer.paypal.com/docs/nvp-soap-api/do-reference-transaction-nvp/#payment-details-fields
-   * https://developer.paypal.com/docs/api/payments/v1/#definition-amount
-   */
-  getPayPalAmountStringFromAmountInCents(amountInCents: number): string {
-    if (amountInCents.toString().length > 10) {
-      throw new AmountExceedsPayPalCharLimitError(amountInCents);
-    }
-    // Left pad with zeros if necessary, so we always get a minimum of 0.01.
-    const amountAsString = String(amountInCents).padStart(3, '0');
-    const dollars = amountAsString.slice(0, -2);
-    const cents = amountAsString.slice(-2);
-    return `${dollars}.${cents}`;
   }
 }
